Settings sections refactor
